Unify look of pages
