+++ frontend/src/components/AdminView.scss
@@ -8,7 +8,7 @@
 
 .title {
     font-weight: 600;
-    margin-top: 60px;
+    margin-top: 73px;
 }
 
 .itemsContainer {
@@ -32,7 +32,7 @@
 
 .searchBy {
     @extend .button;
-    margin-top: 76px;
+    margin-top: 83px;
 
     .MuiSvgIcon-root {
         left: 1.875rem;
@@ -69,7 +69,7 @@
 }
 
 .listItems {
-    padding-top: 83px;
+    padding-top: 1000px;
 }
 
 .listItem {
