+++ frontend/src/components/NavBar.scss
@@ -56,6 +56,10 @@
     cursor: pointer;
 }
 
+.login:hover {
+    text-decoration: underline;
+}
+
 .profile {
     position: absolute;
