+++ frontend/src/components/PollAnswering.scss
@@ -15,7 +15,7 @@
 }
 
 .questionType {
-    margin-top: 100;
+    margin-top: -7px;
     padding-top: 80px;
     font-weight: 600;
 }
